refactor: store 작성 any type으로 지정 - 추후 수정 예정
